move math vector defs to cpp
